refactor: excluindo arquivos individuais modPlano para unificar em um único arquivo.
