--- src/modPlano.h
@@ -1,6 +1,0 @@
-#ifndef MODULO_PLANO_H
-#define MODULO_PLANO_H
-
-char moduloPlano(void);
-
-#endif
