--- src/ui/plano/atualizarPlano.c
@@ -1,18 +1,0 @@
-#include <stdio.h>
-
-char telaAtualizarPlano(void) {
-
-    char op;
-
-    printf("====================\n");
-    printf("ATUALIZAR PLANO\n");
-    printf("====================\n");
-    printf("\n");
-    printf("<Aperte 0 para sair!>\n");
-    
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-
-    }
