--- src/ui/plano/atualizarPlano.h
@@ -1,6 +1,0 @@
-#ifndef ATUALIZAR_PLANO_H
-#define ATUALIZAR_PLANO_H
-
-char telaAtualizarPlano(void);
-
-#endif
