--- src/ui/plano/cadastrarPlano.c
@@ -1,19 +1,0 @@
-#include <stdio.h>
-
-char telaCadastrarPlano(void) {
-    
-    char op;
-
-    printf("====================\n");
-    printf("CADASTRAR PLANO\n");
-    printf("====================\n");
-  
-    printf("  ...EM BREVE... \n");
-    printf("\n");
-    printf("<Aperte 0 para sair!>\n");
-
-    scanf("%c", &op);
-    getchar();
-
-    return op;
-    }
