--- src/ui/plano/cadastrarPlano.h
@@ -1,6 +1,0 @@
-#ifndef CADASTRAR_PLANO_H
-#define CADASTRAR_PLANO_H
-
-char telaCadastrarPlano(void);
-
-#endif
