--- src/ui/plano/excluirPlano.h
@@ -1,6 +1,0 @@
-#ifndef EXCLUIR_PLANO_H
-#define EXCLUIR_PLANO_H
-
-char telaExcluirPlano(void);
-
-#endif
