--- src/ui/plano/telaPlano.h
@@ -1,6 +1,0 @@
-#ifndef TELA_PLANO_H
-#define TELA_PLANO_H
-
-char telaPlano(void);
-
-#endif
