--- src/ui/plano/visualizarPlano.h
@@ -1,6 +1,0 @@
-#ifndef VISUALIZAR_PLANO_H
-#define VISUALIZAR_PLANO_H
-
-char telaVisualizarPlano(void);
-
-#endif
