+++ src/ui/telasPlano.h
@@ -1,0 +1,11 @@
+#ifndef TELASPLANO_H
+#define TELASPLANO_H
+
+char moduloPlano(void);
+char telaPlano(void);
+char telaCadastrarPlano(void);
+char telaVisualizarPlano(void);
+char telaAtualizarPlano(void);
+char telaExcluirPlano(void);
+
+#endif
